backend/pyproject.toml
[project]
name = "firefly-tut"
version = "0.1.0"
description = "Firefly Tutorial"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests==2.32.5",
    "python-dotenv>=1.2.1",
    "fastapi[standard]>=0.120.2",
    "pillow>=12.0.0",
    "boto3>=1.41.1",
    "uvicorn>=0.38.0",
]
